+++ SaveLoope/Core/Models/TransactionRecord.swift
@@ -9,15 +9,15 @@ enum TransactionType: String, Codable {
 @Model
 final class TransactionRecord {
     var id: UUID
-    var amount: Int
+    var amount: Double
     var date: Date
     var type: TransactionType
     var envelope: Envelope?
     var note: String
     var isRecurring: Bool
     var parentId: UUID?
-    
-    init(amount: Int, date: Date, type: TransactionType, envelope: Envelope? = nil, note: String = "", isRecurring: Bool = false, parentId: UUID? = nil) {
+
+    init(amount: Double, date: Date, type: TransactionType, envelope: Envelope? = nil, note: String = "", isRecurring: Bool = false, parentId: UUID? = nil) {
         self.id = UUID()
         self.amount = amount
         self.date = date
